+++ lib/base_decorator.rb
