+++ lib/book_factory.rb
@@ -1,0 +1,7 @@
+require_relative 'book'
+
+class BookFactory
+    def create_book(title, author)
+      Book.new(title, author)
+    end
+  end
