+++ lib/capitalize_decorator.rb
