+++ lib/trim_decorator.rb
